Add domain.model json_schema adapter module
