+++ tests/integration/test_json_schema.py
@@ -1,7 +1,8 @@
-from switchbot.domain.model import SwitchBotDevice, SwitchBotStatus
+from switchbot.domain.model import SwitchBotDevice, SwitchBotStatus, SwitchBotScene
+from switchbot.adapters.json_schema import SwitchBotDeviceSchema, SwitchBotStatusSchema, SwitchBotSceneSchema
 
 
-def test_load_switchbot_device_from_json():
+def test_switchbot_device_schema():
     data = {
         'deviceId': '6055F92FCFD2',
         'deviceName': '小風扇開關',
@@ -9,17 +10,18 @@ def test_load_switchbot_device_from_json():
         'enableCloudService': True,
         'hubDeviceId': ''
     }
-    obj = SwitchBotDevice.fromdict(data)
+    _schema = SwitchBotDeviceSchema()
+    obj = _schema.load(data)
     assert isinstance(obj, SwitchBotDevice)
     assert obj.device_id == data.get('deviceId')
     assert obj.device_name == data.get('deviceName')
     assert obj.device_type == data.get('deviceType')
     assert obj.enable_cloud_service == data.get('enableCloudService')
     assert obj.hub_device_id == data.get('hubDeviceId')
-    assert obj.asdict() == data
+    assert _schema.dump(obj) == data
 
 
-def test_switchbot_state_model():
+def test_switchbot_status_schema():
     data = {
         "deviceId": "6055F92FCFD2",
         "deviceType": "Plug Mini (US)",
@@ -31,7 +33,8 @@ def test_switchbot_state_model():
         "electricCurrent": 0,
         "version": "V1.4-1.4"
     }
-    obj = SwitchBotStatus.fromdict(data)
+    _schema = SwitchBotStatusSchema()
+    obj = _schema.load(data)
     assert isinstance(obj, SwitchBotStatus)
     assert obj.device_id == data.get('deviceId')
     assert obj.device_type == data.get('deviceType')
@@ -42,4 +45,18 @@ def test_switchbot_state_model():
     assert obj.electricity_of_day == data.get('electricityOfDay')
     assert obj.electric_current == data.get('electricCurrent')
     assert obj.version == data.get('version')
-    assert obj.asdict() == data
+    assert _schema.dump(obj) == data
+
+
+def test_switchbot_scene_schema():
+    data = {
+        "sceneId": "T01-202309291436-01716250",
+        "sceneName": "allOff"
+    }
+    _schema = SwitchBotSceneSchema()
+    obj = _schema.load(data)
+    assert isinstance(obj, SwitchBotScene)
+    assert obj.scene_id == data.get('sceneId')
+    assert obj.scene_name == data.get('sceneName')
+    assert _schema.dump(obj) == data
+
